% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipa.R
\name{add_alpha}
\alias{add_alpha}
\title{Add transparency (set pixels to zero) area to image}
\usage{
add_alpha(img, area, quiet = TRUE)
}
\arguments{
\item{img}{image object (\code{cimg} class)}

\item{area}{area to modify: \code{c(x0, width, y0, height)}
    where: \code{x0} is the starting pixel on the x-axis,
           \code{width} is the number of pixels along x,
           \code{y0} is the starting pixel on the y-axis, and
           \code{height} is the number of pixels along y

    To invert the area (trim right to left or bottom to top), set
    \code{x0 = -1} or \code{y0 = -1} respectively.

    To modify from \code{x0} to full width, set \code{width = -1}. Similarly,
    for the full height, set \code{height = -1}.}

\item{quiet}{boolean flag to show message of work area}
}
\value{
modified image object (\code{cimg} class)
}
\description{
Add transparency (set pixels to zero) area to image
}
\examples{
\dontrun{
    # Create test image
    red <- matrix(0, 50, 50)
    red[1:25, 1:25] <- 1
    blue <- matrix(0, 50, 50)
    blue[26:50, 1:25] <- 1
    green <- matrix(0, 50, 50)
    green[1:25, 26:50] <- 1
    alpha <- matrix(1, 50, 50)
    alpha[26:50, 26:50] <- 0
    img <- imager::as.cimg(abind::abind(imager::as.cimg(red),
                                        imager::as.cimg(blue),
                                        imager::as.cimg(green),
                                        imager::as.cimg(alpha),
                                        along = 4))
    # Remove red portion of the image
    plot(add_alpha(img, c(1, 25, 1, 25)))
    # Remove red and green portions
    plot(add_alpha(img, c(1, -1, 1, 25)))
    # Remove red and blue portions
    plot(add_alpha(img, c(1, 25, 1, -1)))
    # Remove green and alpha portions
    plot(add_alpha(img, c(-1, 25, 1, -1)))
}
}
