% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adductsIPA.R
\docType{data}
\name{adductsIPA}
\alias{adductsIPA}
\title{adductsIPA}
\format{A data frame with 59 rows and 9 variables}
\usage{
data(adductsIPA)
}
\description{
A dataframe containing the information need to compute all the possible adducts
}
\keyword{datasets}
