% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_adducts_neg.R
\docType{data}
\name{all_adducts_NEG}
\alias{all_adducts_NEG}
\title{all_adducts_NEG}
\format{A data frame with 62558 rows and 7 variables}
\usage{
data(all_adducts_NEG)
}
\description{
A dataframe containing all the possible adducts for negative ionisation mode
}
\keyword{datasets}
