% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_adducts_neg_updated.R
\docType{data}
\name{all_adducts_NEG_updated}
\alias{all_adducts_NEG_updated}
\title{all_adducts_NEG_updated @aliases}
\format{A data frame with 62558 rows and 7 variables}
\usage{
data(all_adducts_NEG_updated)
}
\description{
A dataframe containing all the possible adducts for negative ionisation mode
}
\keyword{datasets}
