% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_adducts_pos_updated.R
\docType{data}
\name{all_adducts_POS_updated}
\alias{all_adducts_POS_updated}
\title{all_adducts_POS_updated}
\format{A data frame with 62828 rows and 7 variables}
\usage{
data(all_adducts_POS_updated)
}
\description{
A dataframe containing all the possible adducts for positive ionisation mode
}
\keyword{datasets}
