% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{as_data_stack}
\alias{as_data_stack}
\alias{as_data_list}
\title{Prepare data stacks/lists}
\usage{
as_data_stack(data)

as_data_list(data)
}
\arguments{
\item{data}{This argument should be a list a list of
imputed dataframes, each having the same number of rows and columns.}
}
\description{
stacked dataframes comprising multiply imputed
dataframes stacked on top of one another. Imputed values in
these objects are usually a model prediction + random noise.
The additional noise creates diversity among the multiple
dataframes. Data lists are similar, but data frames are kept
separate and one model is fit for each data set.

These functions create ID variables and attach classes and attributes
that are used to direct downstream analyses.
}
\examples{
data_list <- list(
  df1 = data.frame(a = 1:3, b = letters[1:3]),
  df2 = data.frame(a = 2:4, b = letters[2:4])
)

as_data_stack(data_list)

as_mi(data_list)

as_si(data_list[[1]])
}
