% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{as_midy}
\alias{as_midy}
\alias{as_mi}
\title{Prepare \code{si}/\code{mi}/\code{midy} data}
\usage{
as_midy(data)

as_mi(data)
}
\arguments{
\item{data}{This argument should be a list a list of
imputed dataframes, each having the same number of rows and columns,
for \code{as_midy} and \code{as_mi}. For \code{as_si}, this argument should be
a data frame with imputed values.}
}
\description{
\code{si} objects are singly imputed dataframes where
missing values are filled in with a model's predicted value.
\code{mi} and \code{midy} objects are dataframes comprising multiply imputed
dataframes stacked on top of one another. Imputed values in
\code{mi} and \code{midy} objects are a model prediction + random noise.
The additional noise creates diversity among the multiple
dataframes. Data in \code{mi} objects are sorted by \code{._mi.ID_.},
which identifies dataframes. Data in \code{midy} objects are sorted
by \code{._midy.ID_.}, which identifies rows in the original (unimputed)
data.

These functions create ID variables and attach classes and attributes
that are used to direct downstream analyses.
}
\note{
Imputed dataframes are copies of an original dataframe,
with cells that originally had missing values filled in
with a plausible estimate.
}
\examples{
data_list <- list(
  df1 = data.frame(a = 1:3, b = letters[1:3]),
  df2 = data.frame(a = 2:4, b = letters[2:4])
)

as_midy(data_list)

as_mi(data_list)

as_si(data_list[[1]])
}
