% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{bottle}
\alias{bottle}
\alias{bottle.soft_brew}
\title{Bottle a brew}
\usage{
bottle(brew, flavor = c("tibble", "matrix"))

\method{bottle}{soft_brew}(brew, flavor = c("tibble", "matrix"))
}
\arguments{
\item{brew}{a \code{brew} object that has been passed through all
the brewing steps: \link{malt}, \link{mash}, \link{boil}, and \link{ferment}.}

\item{flavor}{a character value indicating what composition
the training and testing data will be returned as.}
}
\description{
Bottle a brew
}
\section{Methods (by class)}{
\itemize{
\item \code{soft_brew}: bottle up a soft brew
}}

