% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Build_add_matrix.R
\name{build.add.connenctivity.matrix}
\alias{build.add.connenctivity.matrix}
\title{Building Adducts connectivity matrix}
\usage{
build.add.connenctivity.matrix(Prior, DB, ionisation,
  fully.connected = FALSE, v = TRUE, IT = 500)
}
\arguments{
\item{Prior}{The output of compute.Priors() function}

\item{DB}{The database used in a dataframe format. It has to be organized in the same fashion of the one included with the package}

\item{ionisation}{A character indicating the ionisation mode of the experiment, either 'positive' or 'negative'}

\item{fully.connected}{A logical value if TRUE all adducts will be considered connected with each other, if FALSE (default) all
the adducts are only connected to the main adduct}

\item{v}{A logical value indicating if the progress will be shown (default TRUE)}

\item{IT}{A number inticating after how many iteration an update should be shown (default 500)}
}
\value{
A binary matrix encoding the adducts connections
}
\description{
This functions created the Adducts connectivity matrix needed for the posterior estimation
}
\seealso{
build.bio.connectivity.matrix build.iso.connectivity.matrix
}
\author{
Francesco Del Carratore \email{francescodc87@gmail.com}
}
