% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Build_bio_matrix.R
\name{build.bio.connectivity.matrix}
\alias{build.bio.connectivity.matrix}
\title{Building Biotransformations connectivity matrix}
\usage{
build.bio.connectivity.matrix(Prior, DB, ionisation,
  connection.type = "reactions", connections = NULL, v = TRUE,
  IT = 500)
}
\arguments{
\item{Prior}{The output of compute.Priors() function}

\item{DB}{The database used in a dataframe format. It has to be organized in the same fashion of the one included with the package}

\item{ionisation}{A character indicating the ionisation mode of the experiment, either 'positive' or 'negative'}

\item{connection.type}{if equal to "reactions" (default) considers the reactions reported in the database. If it is equal to "formulas", the function consider the list of possible reactions reported in connections}

\item{connections}{if connection.type=="fomulas", this vector contains a list of formulas that are likely to be added or substracted by an enzymatic reaction. If it is equal to NULL the default list of formulas is used}

\item{v}{A logical value indicating if the progress will be shown (default TRUE)}

\item{IT}{A number inticating after how many iteration an update should be shown (default 500)}
}
\value{
A binary matrix encoding the biotransfiormations connections
}
\description{
This functions created the Biotransformations connectivity matrix needed for the posterior estimation
}
\seealso{
build.add.connectivity.matrix build.iso.connectivity.matrix
}
\author{
Francesco Del Carratore \email{francescodc87@gmail.com}
}
