% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Build_iso_matrix.R
\name{build.iso.connectivity.matrix}
\alias{build.iso.connectivity.matrix}
\title{Building Isotopes connectivity matrix}
\usage{
build.iso.connectivity.matrix(Prior, DB, ratios = TRUE, v = TRUE,
  IT = 500)
}
\arguments{
\item{Prior}{The output of compute.Priors() function}

\item{DB}{The database used in a dataframe format. It has to be organized in the same fashion of the one included with the package}

\item{ratios}{A logical value indicating if the matrix should contain the expected intensity ratios (default TRUE)}

\item{v}{A logical value indicating if the progress will be shown (default TRUE)}

\item{IT}{A number inticating after how many iteration an update should be shown (default 500)}
}
\value{
A matrix encoding the isotopes connections
}
\description{
This functions created the Isotopes connectivity matrix needed for the posterior estimation
}
\seealso{
build.add.connectivity.matrix build.bio.connectivity.matrix
}
\author{
Francesco Del Carratore \email{francescodc87@gmail.com}
}
