% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computing_priors.R
\name{compute.Priors}
\alias{compute.Priors}
\title{Computing prior probabilities associated with putative annotations}
\usage{
compute.Priors(Hits, dataset, pk = rep(1, nrow(Hits$all.formulas)), ppm,
  unknown.ppm = NA, RT.pen = 0.5, v = T, IT = 120,
  pr.lim = 1e-05)
}
\arguments{
\item{Hits}{The output of find.hits() function}

\item{dataset}{A matrix containing the measured data, organized in 3 colums: mz, RT and Int}

\item{pk}{A vector of length nrow(Hits$all.formulas). Values between 0 and 1 expressing initial confidence of the presence of each formula}

\item{ppm}{A number indicating the instrument accuracy to be considered}

\item{unknown.ppm}{The ppm number to be assigned to unknown (default NA - No unknown is considered)}

\item{RT.pen}{A numerical value indicating the multiplicative factor used when the measured retention time is outside of the range reported in the database}

\item{v}{A logical value indicating if the progress will be shown (default TRUE)}

\item{IT}{A number inticating after how many iteration an update should be shown (default 120)}

\item{pr.lim}{A number inidicating the lowest probability value allowed (default 1e-05)}
}
\value{
A list containing the matrix of the prior probabilities and the id.masses and all.formulas objects
}
\description{
This functions takes as input the output of the find.hits() function and computes the prior
probabilities for the putative annotations
}
\seealso{
find.Hits IPAposteriors
}
\author{
Francesco Del Carratore \email{francescodc87@gmail.com}
}
