% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_updated.R
\docType{data}
\name{DB_updated}
\alias{DB_updated}
\title{Updated Database}
\format{A data frame with 15707 rows and 14 variables}
\usage{
data(DB_updated)
}
\description{
Updated Database for the IPA package
}
\keyword{datasets}
