% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diabetes}
\alias{diabetes}
\title{Diabetes data.}
\format{A list with two data frames: one holding complete data and the
other holding data that are missing completely at random. Each set
contains the same participants and the same variables:
\itemize{
\item \code{chol}: Total cholesterol
\item \code{stable_glucose}: Stabilized glucose
\item \code{hdl}: High density lipoprotein
\item \code{age}: age, years
\item \code{gender_male}: male gender
\item \code{frame_medium}: medium frame (body type) indicator
\item \code{frame_small}: small frame (body type) indicator
\item \code{sbp}: systolic blood pressure, mm Hg (1 measurement)
\item \code{dbp}: diastolic blood pressure, mm Hg (one measurement)
\item \code{waist}: waist circumference, inches
\item \code{hip}: hip circumference, inches
\item \code{time_ppn}: Postprandial time when labs were drawn in minutes
\item \code{height_cm}: height, cm
\item \code{weight_kg}: weight, kg
\item \code{diabetes}: diabetes status (Yes or No)
}}
\source{
\url{http://192.38.117.59/~tag/Teaching/share/data/Diabetes.html}
}
\usage{
diabetes
}
\description{
These data are courtesy of Dr John Schorling, Department of Medicine,
University of Virginia School of Medicine. The data originally described
1046 participants who were interviewed in a study to understand the prevalence
of obesity, diabetes, and other cardiovascular risk factors in central
Virginia for African Americans. A total of 403 participants were screened for
diabetes, which was defined as glycosolated hemoglobin > 7.0. Among those
403 participants, 366 provided complete data for the variables in these data.
}
\details{
The data were originally shared in the \code{Publish} package, developed by
Thomas Gerds. They are modified slightly here for the purposes of the
\code{ipa} package.
}
\references{
Willems JP, Saunders JT, DE Hunt, JB Schorling: Prevalence of coronary
heart disease risk factors among rural blacks: A community-based study.
Southern Medical Journal 90:814-820; 1997 Schorling JB, Roach J, Siegel M,
Baturka N, Hunt DE, Guterbock TM, Stewart HL: A trial of church-based
smoking cessation interventions for rural African Americans. Preventive
Medicine 26:92-101; 1997.
}
\keyword{datasets}
