% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drop_extension}
\alias{drop_extension}
\title{Drop the file extension in a string}
\usage{
drop_extension(filename, only_basename = FALSE)
}
\arguments{
\item{filename}{filename w/o full path}

\item{only_basename}{boolean flag to indicate if only the base name should
be returned.}
}
\value{
filename without extension
}
\description{
Drop the file extension in a string
}
\examples{
drop_extension("/path/to/A.txt")
drop_extension("/path/to/A.txt", TRUE)
}
