% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_na.R
\name{fill_na}
\alias{fill_na}
\title{Fill missing values}
\usage{
fill_na(data, vals, na_indx = NULL, make_copy = TRUE)
}
\arguments{
\item{data}{a data frame with missing values}

\item{vals}{a named list. Names correspond to columns in \code{data} with
missing values.}

\item{na_indx}{a named list containing the indices where missing data
are located. Names should correspond to columns in \code{data}. The
\link{mindx} function will create this type of object.}

\item{make_copy}{a logical value that is only relevant if \code{data}
is a \code{data.table}. if \code{TRUE}, a new object is created and \code{data}
will not be modified. If \code{FALSE}, \code{data} will be modified in place.
The latter will be more efficient but is onyl feasible when you
only need to make one imputed data frame.}
}
\value{
a data frame with missing values filled in.
}
\description{
Fill missing values
}
\examples{

df = data.frame(A = 1:5, B = letters[1:5])
df[1, 1] = NA
df[2, 2] = NA

vals = list(A = 111, B = 'bbb')
fill_na(df, vals)

}
