% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_hits.R
\name{find.hits}
\alias{find.hits}
\title{Finding putative annotations}
\usage{
find.hits(adducts.matrix, dataset, ppm.thr, RTwin, isotopes,
  iso.threshold = 1, corr.matrix = NULL, corr.thr = 0.75,
  relation.id = NULL, v = T, IT = 500)
}
\arguments{
\item{adducts.matrix}{A matrix contains information about all possible adducts.
Columns are: KEGG.id (e.g. 'C00002'), adduct (e.g. 'M+H'), RT (previously known
retention time for the compound, NA if unknown), formula (e.g. C10H17N5O13P3),
theoretical mz (e.g. 508.003570124), charge (e.g. -1) mono ('mono'
if monoisotopical form, 'iso' otherwise)}

\item{dataset}{A matrix containing the measured data, organized in 3 colums: mz, RT and Int}

\item{ppm.thr}{A numerical value indicating the maximum accuracy value to be considered}

\item{RTwin}{A numerical value indicating the maximum difference allowed
between measured RT of different isotopes (should be significantly lower than RTwin1)}

\item{isotopes}{A matrix containing infomation about isotopes}

\item{iso.threshold}{A numerical value indicating the probability below which
isotope peaks can be omitted}

\item{corr.matrix}{A matrix containing the correlation values between peaks}

\item{corr.thr}{A numerical value expressing the treshold used to consider the correlations significant}

\item{relation.id}{A vector containg class labels of the previously grouped peaks}

\item{v}{A logical value indicating if the progress will be shown (default TRUE)}

\item{IT}{A number inticating after how many iteration an update should be shown (default 120)}
}
\value{
A list containing the matrix of the posterior probabilities, the id.masses vector, the all.formulas dataframe and
the allsampcomp matrix containing all the assignments for each iteration of the Gibbs sampler
}
\description{
This function finds all the possible annotations given an user-defined ppm window.
}
\seealso{
find.Hits IPAposteriors
}
\author{
Francesco Del Carratore \email{francescodc87@gmail.com}
}
