% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{gen_simdata}
\alias{gen_simdata}
\title{a simple framework to simulate simple dataframes}
\usage{
gen_simdata(problem_type = "regression", ncov = 50, nint = 20,
  rho = 1/2, nobs = 10000, error_sd = 1/2, prevalence = NULL,
  split_prop = 1/2, miss_pattern = "mar", trn_miss_prop = 1/2,
  tst_miss_prop = 0)
}
\arguments{
\item{problem_type}{A character value indicating the problem type to
simulate data for. Valid options are 'regression', 'classification',
or 'survival'.}

\item{ncov}{the number of main effects used to generate an outcome variable}

\item{nint}{the number of interaction effects used to generate an outcome
variable.}

\item{rho}{the auto correlation constant among predictors in the X matrix.}

\item{nobs}{the total number of observations in the simulated data.}

\item{error_sd}{the standard deviation of error applied when generating
outcome values.}

\item{prevalence}{(this is only relevant for classification problems).
The prevalence of the outcome.}

\item{split_prop}{the proportion of data that will be randomly assigned
to the training dataset.}

\item{miss_pattern}{A character value indicating what type of missingness
pattern to apply. Valid options are 'mcar', 'mar', and 'mnar'.
mcar = missing completely at random, mar = missing at random, and
mnar = missing not at random.}

\item{trn_miss_prop}{The proportion of data in the training set that will
be set to missing.}

\item{tst_miss_prop}{The proportion of data in the testing set that will
be set to missing.}
}
\description{
a simple framework to simulate simple dataframes
}
