% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_it.R
\name{get_factor_levels}
\alias{get_factor_levels}
\title{get factor levels from data}
\usage{
get_factor_levels(data)
}
\arguments{
\item{data}{data with factor variables}
}
\description{
get factor levels from data
}
