% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_net.R
\name{impute_net}
\alias{impute_net}
\title{Elastic net imputation}
\usage{
impute_net(
  data_ref,
  data_new = NULL,
  fit = NULL,
  cols = dplyr::everything(),
  df_min = 1,
  df_max = 10,
  df_stp = 1,
  restore_data = TRUE,
  verbose = 1
)
}
\arguments{
\item{data_ref}{a data frame.}

\item{data_new}{an optional data frame. If supplied, then \code{data_ref}
will be used as a reference dataset for \code{data_new} and the output
will contain imputed values for \code{data_new}. If not supplied,
the output will contain imputed values for \code{data_ref}.}

\item{fit}{a list of lists of strings that contain formulas
from previously fitted imputation models. This input variable
is created using \code{data_ref} when \code{data_new = NULL}.}

\item{cols}{columns that should be imputed and/or used to impute other
columns. Supports tidy select functions (see examples).}

\item{df_min, df_max}{integer value designating the minimum
and maximum degrees of freedom in penalized regression models.}

\item{df_stp}{integer value indicating step size for model
degrees of freedom between successive imputations.}

\item{restore_data}{a logical value. If \code{TRUE}, the variable types
of the imputed values will match those of the original data. If \code{FALSE},
the imputed values are returned in a one-hot encoded format.}

\item{verbose}{an integer value of 0, 1, or 2. If \code{verbose = 0}, nothing
is printed. If \code{verbose = 1}, messages are printed to the console showing
what general steps are being taken in the imputation process. If
\code{verbose = 2}, all relevant information on convergence is printed in
addition to general messages.}
}
\description{
\code{glmnet} is used to impute missing values.
For more details, see \link[glmnet:glmnet]{glmnet}
}
