% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipa-package.R
\docType{package}
\name{ipa-package}
\alias{ipa}
\alias{ipa-package}
\title{ipa: Convert Between Phonetic Alphabets}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Converts character vectors between phonetic
    representations.  Supports IPA (International Phonetic Alphabet),
    X-SAMPA (Extended Speech Assessment Methods Phonetic Alphabet), and
    ARPABET (used by the CMU Pronouncing Dictionary).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rossellhayes/ipa}
  \item Report bugs at \url{https://github.com/rossellhayes/ipa/issues}
}

}
\author{
\strong{Maintainer}: Alexander Rossell Hayes \email{alexander@rossellhayes.com} (\href{https://orcid.org/0000-0001-9412-0457}{ORCID}) [copyright holder]

}
\keyword{internal}
