% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_functions.R
\name{is_brew}
\alias{is_brew}
\alias{is_spiced}
\alias{is_mashed}
\alias{is_fermented}
\alias{is_bottled}
\title{Brew types}
\usage{
is_brew(x)

is_spiced(x)

is_mashed(x)

is_fermented(x)

is_bottled(x)
}
\arguments{
\item{x}{a brew object}
}
\value{
a logical value
}
\description{
Functions to help identify what stage a brew is in.
}
