% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{knn_fill}
\alias{knn_fill}
\title{k-nearest-neighbor (knn) imputation}
\usage{
knn_fill(data, new_data, outcome, n_impute = 10, step_size = 1,
  neighbor_sequence = NULL, verbose = TRUE, ...)
}
\description{
k-nearest-neighbor (knn) imputation
}
