% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{knn_fit}
\alias{knn_fit}
\title{k-nearest-neighbor (knn) imputation}
\usage{
knn_fit(data, outcome, n_impute = 10, step_size = 1,
  neighbor_sequence = NULL, verbose = TRUE, ...)
}
\description{
Adapted from Max Kuhn's knn imputation functions in recipes
}
