% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_label.R
\name{stack_label}
\alias{stack_label}
\alias{list_label}
\title{Stack labels}
\usage{
stack_label(label, n_impute)

list_label(label, n_impute)
}
\arguments{
\item{label}{A vector of label values. For continuous outcomes, the
vector should have numeric values. For categorical outcomes, the
vector should be a factor. For survival outcomes, the vector
should be numeric and contain values indicating time until
the event or censoring (censored values should be negative).}

\item{n_impute}{An integer indicating the number of imputed datasets.}
}
\description{
When \code{xgboost} is applied to lists of datasets or stacked data
(i.e., multiply imputed data ),
the label vector must be adjusted so that it is consistent with
the structure of the x matrix (or matrices). This function
is handy for expanding label vectors in this manner.
}
\details{
\code{stack_label()}can be used to create vectors of labels, and
\code{list_label()} creates lists of vectors of labels.
}
\examples{

y <- c(1,2,3)

stack_label(y, n_impute = 2)
list_label(y, n_impute = 2)

}
