% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{malt_nbrs}
\alias{malt_nbrs}
\title{Make a neighborhood malt.}
\usage{
malt_nbrs(training_data, outcome)
}
\arguments{
\item{training_data}{data that will be used to create imputation models}

\item{outcome}{column name(s) of outcomes in \code{training_data}.}
}
\description{
This set of functions is an adaptation of Max Kuhn's
nearest neighbor imputation functions in the \link[recipes:recipes]{recipes}
and \code{caret} packages. It also uses the \code{gower} package to implement
algorithms that compute Gower's distance.

What makes this type of nearest neighbor imputation
different is its flexibility in the number of neighbors used
to impute missing values and the aggregation function applied.
For example, to create 10 imputed datasets that use 1, 2, ..., 10
neighbors to impute missing values would require fitting
10 separate nearest neighbors models using the conventional functions.
The \code{ipa} package lets a user create all of these imputed sets
with just one fitting of a nearest neighbor model. Additionally,
for users who want to use nearest neighbors for multiple imputation,
\code{ipa} gives the option to sample 1 neighbor value at random from
a neighborhood, rather than aggregate values into a summary. This
sampling strategy will increase the diversity of imputed datasets.
}
\note{
Gower (1971) originally defined a similarity measure (s, say) with
values ranging from 0 (completely dissimilar) to 1 (completely similar).
The distance returned here equals 1-s.
}
\section{Overview}{

Brewing your own IPA is a lot like handling missing
data. When a brew is initiated, the first step is to determine
the type of wheat and prepare it for mashing. For the \code{ipa}
package, this amounts to determining the type of imputation
method and preparing it for use. The \code{brew} object is
prepared for mashing (the next step) by identifying a
training data set and the outcome(s) of the analysis.
}

\references{
Gower, John C.
"A general coefficient of similarity and some of its properties."
Biometrics (1971): 857-871.
}
