% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{mash_nbrs}
\alias{mash_nbrs}
\title{Make a neighborhood mash}
\usage{
mash_nbrs(malt, n_impute = 10L, step_size = NULL, min_neighbors = 1L,
  neighborhood = NULL)
}
\description{
Make a neighborhood mash
}
