% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mash.R
\name{masher_rngr}
\alias{masher_rngr}
\title{Ranger's masher}
\usage{
masher_rngr(maxiter = 10L, num.trees = 100, sample.fraction = 0.632)
}
\arguments{
\item{maxiter}{An integer specifying the maximum number of
iterations in the \link[missRanger:missRanger]{missRanger} algorithm.}

\item{num.trees}{An integer specifying the number of decision
trees fitted for each random forest. Lower values of \code{num.trees}
will increase computation speed but may decrease accuracy.}

\item{sample.fraction}{A fraction indicating the proportion of
data randomly sampled for each decision tree. Lower values of
\code{sample.fraction} will increase computation speed but may
decrease accuracy.}
}
\value{
a list with \link[missRanger:missRanger]{missRanger} and
\link[ranger:ranger]{ranger} input values that can be passed
directly into \code{missRanger_brew} objects via \link{mash}.
}
\description{
It can be a little overwhelming to remember which sets of
parameters go with each \code{ipa_brew} flavor, so just pair your
flavor with its \code{masher} function and get on with your \code{brew}.
}
