% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mash.R
\name{masher_soft}
\alias{masher_soft}
\title{Soft masher}
\usage{
masher_soft(
  bs = TRUE,
  bs_maxit = 20,
  bs_thresh = 1e-09,
  bs_row.center = FALSE,
  bs_col.center = TRUE,
  bs_row.scale = FALSE,
  bs_col.scale = TRUE,
  si_type = "als",
  si_thresh = 1e-05,
  si_maxit = 100,
  si_final.svd = TRUE
)
}
\arguments{
\item{bs}{a logical value. If \code{TRUE}, \code{\link[softImpute:biScale]{softImpute::biScale()}} is applied
to \code{data_ref} or \code{rbind(data_ref, data_new)} prior to fitting \code{softImpute}
models.}

\item{bs_maxit}{an integer indicating the maximum number of iterations
for the \code{biScale} algorithm.}

\item{bs_thresh}{convergence threshold for the \code{biScale} algorithm.}

\item{bs_row.center}{a logical value. If \code{TRUE}, row centering
will be performed. If \code{FALSE} (default), then nothing is done.}

\item{bs_col.center}{a logical value. If \code{TRUE} (default), column centering
will be performed. If \code{FALSE}, then nothing is done.}

\item{bs_row.scale}{a logical value. If \code{TRUE}, row scaling
will be performed. If \code{FALSE} (default), then nothing is done.}

\item{bs_col.scale}{a logical value. If \code{TRUE} (default), column scaling
will be performed. If \code{FALSE}, then nothing is done.}

\item{si_type}{two algorithms are implemented, type="svd" or the default
type="als". The "svd" algorithm repeatedly computes the svd of the
completed matrix, and soft thresholds its singular values. Each new
soft-thresholded svd is used to re-impute the missing entries. For
large matrices of class "Incomplete", the svd is achieved by an
efficient form of alternating orthogonal ridge regression. The
"als" algorithm uses this same alternating ridge regression, but
updates the imputation at each step, leading to quite substantial
speedups in some cases. The "als" approach does not currently
have the same theoretical convergence guarantees as the
"svd" approach.}

\item{si_thresh}{convergence threshold for the \code{softImpute} algorithm,
measured as the relative change in the Frobenius norm between two
successive estimates.}

\item{si_maxit}{maximum number of iterations for the \code{softImpute}
algorithm.}

\item{si_final.svd}{only applicable to \code{si_type = "als"}. The alternating
ridge-regressions do not lead to exact zeros. With the default
\code{final.svd = TRUE}, at the final iteration, a one step unregularized
iteration is performed, followed by soft-thresholding of the
singular values, leading to hard zeros.}
}
\description{
It can be a little overwhelming to remember which sets of
parameters go with each \code{ipa_brew} flavor. If you pair your
flavor with its \code{masher} function, e.g.
\verb{mash(brew, with = masher_<flavor>())}, you can use tab-completion
inside of \verb{masher_<flavor>()} to see which arguments should be
specified for your mash.
}
