% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgboost.R
\name{mgb_bhaz}
\alias{mgb_bhaz}
\title{Baseline hazard function}
\usage{
mgb_bhaz(mgb_booster, eval_times = NULL, smooth = FALSE,
  cumulative = TRUE)
}
\arguments{
\item{eval_times}{Values at which the baseline hazard will
be evaluated.}

\item{smooth}{If \code{TRUE} \code{mgb_bhaz} will smooth the estimated
baseline hazard using Friedman's super smoother
\code{\link[stats:supsmu]{stats::supsmu()}}.}

\item{cumulative}{If \code{TRUE} the cumulative survival function
will be computed.}

\item{label}{A numeric vector with time-to-event values, where
censored observations have negative times and uncensored
observations have positive times (see \code{\link[=label_for_survival]{label_for_survival()}}).}

\item{predictions}{The predicted values of the regression model
on the log hazard scale.}
}
\description{
This function is a wrapper for the
\code{\link[gbm:basehaz.gbm]{gbm::basehaz.gbm()}}. The function
computes the Breslow estimator of the baseline
hazard function for a proportional hazard
regression model.
}
