% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgboost.R
\name{mgb_cv}
\alias{mgb_cv}
\title{(ipa) Cross-validation for xgboosters}
\usage{
mgb_cv(..., miss_strat, n_impute = NULL)
}
\arguments{
\item{...}{other parameters to pass to \code{params}.}
}
\value{
a modified version of the output from
\code{\link[xgboost:xgb.cv]{xgboost::xgb.cv()}}. Specifically,
an additional item, \code{compute_time} is included
in the output list.
}
\description{
This function is a wrapper for the
\code{\link[xgboost:xgb.cv]{xgboost::xgb.cv()}} function that automates some
adjustments that should be applied when using \code{ipa} data.
For example, the \code{subsample} tuning parameter is automatically
adjusted based on the number of imputed datasets for \code{ipa} data.
Additionally, the folds argument is adjusted such that
observations in \code{ipa} data belonging to the same ID are
allocated to the same data-fold.
}
\details{
To ensure valid data-folds, the \code{folds} argument must be
supplied. If \code{nfolds} is supplied instead, an error will occur.
}
