% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midy.R
\docType{package}
\name{midy-package}
\alias{midy}
\alias{midy-package}
\title{midy: a framework for multiple imputation with decision trees
in the tidyverse.}
\description{
midy provides a flexible set of tools to help develop
ensembles of decision trees using training data that contain
missing values.
}
\details{
It has two main goals:

\itemize{
\item Provide functions to conduct multiple imputation in roughly the
same computational cost as single imputation with common algorithms
(i.e., \code{\link[missForest]{missForest}} and \code{\link[VIM]{kNN}}).
\item Provide convenient wrappers for xgboost that allow a simplified
workflow to compare different strategies to handle missing values.
}

To learn more about midy, start with the vignettes:
\code{browseVignettes(package = "midy")}
}
