% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_miss.R
\name{mindx}
\alias{mindx}
\title{Missing index}
\usage{
mindx(data, drop_empty = TRUE)
}
\arguments{
\item{data}{a data set with missing values.}

\item{drop_empty}{a logical value. If \code{TRUE}, columns in \code{data} with no
missing values will be dropped from the output. If \code{FALSE}, all column
names in \code{data} will be present in the output.}
}
\value{
a list with indices of missing values in \code{data}.
}
\description{
Missing index
}
\examples{

dat <- data.frame(a = c(1, NA), b = c(NA, 2), c = c(1,2))

mindx(dat)
mindx(dat, drop_empty = FALSE)


}
