% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_nbrs.R
\name{mode_est}
\alias{mode_est}
\alias{medn_est}
\alias{medn_est_conserve}
\title{Neighbor aggregates}
\usage{
mode_est(x, random_ties = FALSE)

medn_est(x)

medn_est_conserve(x)
}
\arguments{
\item{x}{a vector of character or integer values for \code{mode_est}. For
\code{medn_est}, only integer values are allowed.}

\item{random_ties}{a logical value indicating whether ties should be
broken at random when two values occur at the same frequency in \code{x}.}
}
\value{
an aggregate scalar with the same type as \code{x}.
}
\description{
Neighbor aggregates
}
\examples{

x <- c('a', 'a', 'b')
mode_est(x)

x <- c(1L, 1L, 2L)
medn_est(x)



}
