% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{nbrs_malt}
\alias{nbrs_malt}
\title{Make a neighborhood malt.}
\usage{
nbrs_malt(training_data, outcome)
}
\arguments{
\item{training_data}{data that will be used to create imputation models}

\item{outcome}{column name(s) of outcomes in \code{training_data}.}
}
\description{
Make a neighborhood malt.
}
\section{Overview}{

Brewing your own IPA is a lot like handling missing
data. When a brew is initiated, the first step is to determine
the type of wheat and prepare it for mashing. For the \code{ipa}
package, this amounts to determining the type of imputation
method and preparing it for use. The \code{brew} object is
prepared for mashing (the next step) by identifying a
training data set and the outcome(s) of the analysis.
}

