% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrimp.R
\name{net_args}
\alias{net_args}
\title{glmnet arguments helper function}
\usage{
net_args(
  alpha = 1/2,
  cmplx = "lambda.1se",
  weights = NULL,
  foldid = NULL,
  nfolds = 10,
  keep_mdl = FALSE,
  keep_prd = TRUE
)
}
\arguments{
\item{alpha}{the elastic net mixing parameter}

\item{cmplx}{Value(s) of the penalty parameter lambda at which
predictions are required. Default is the value s = "lambda.1se" stored
on the CV object. Alternatively s = "lambda.min" can be used. If s
is numeric, it is taken as the value(s) of lambda to be used.
(For historical reasons the glmnet authors use the symbol 's' to
reference this parameter)}

\item{weights}{observation weights. Default is 1 for each observation}

\item{foldid}{an optional vector of values between 1 and nfold
identifying what fold each observation is in. If supplied,
nfold can be missing.}

\item{nfolds}{number of folds - default is 10. Although nfolds can be
as large as the sample size (leave-one-out CV), it is not recommended
for large datasets. Smallest value allowable is nfolds=3}

\item{keep_mdl}{a logical value. If \code{TRUE}, then the output of
\link{scrimp_mdl} will include the \code{glmnet} model object. Default is \code{FALSE}}

\item{keep_prd}{a logical value. If \code{TRUE}, then the output of
\link{scrimp_mdl} will include predictions from the \code{glmnet} model object.
Default is \code{TRUE}}
}
\value{
a named list with arguments that should be passed into the
\code{.fun_args} argument of \link{scrimp_mdl}.
}
\description{
glmnet arguments helper function
}
\examples{

net_args()

net_args(cmplx = 'lambda.min')

}
