% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_results.R
\name{ParseIPAresults}
\alias{ParseIPAresults}
\title{Parse IPA resutls}
\usage{
ParseIPAresults(Post, Prior, dataset, DB, IDs = NULL, v = TRUE,
  IT = 500)
}
\arguments{
\item{Post}{The output of IPAposteriors() function}

\item{Prior}{The output of compute.Priors() function}

\item{dataset}{The initial dataset}

\item{DB}{Database}

\item{IDs}{name of each feature in the initial dataset}

\item{v}{A logical value indicating if the progress will be shown (default TRUE)}

\item{IT}{A number inticating after how many iteration an update should be shown (default 120)}
}
\value{
A list containing the parsed results
}
\description{
This functions takes as input the output of the IPAposteriors() function and parses
the results in a more readable fashion
}
\seealso{
find.Hits compute.Priors IPAposteriors
}
\author{
Francesco Del Carratore \email{francescodc87@gmail.com}
}
