% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brew.R
\name{print.ipa_brew}
\alias{print.ipa_brew}
\alias{is_brew}
\title{Print a brew}
\usage{
\method{print}{ipa_brew}(x, ...)

is_brew(x)
}
\arguments{
\item{x}{an \code{ipa_brew} object.}

\item{...}{additional arguments (currently not used)}
}
\description{
Sometimes you need to see the brew.
}
