% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipa.R
\name{rgb_decomposition}
\alias{rgb_decomposition}
\title{Generate an RGB decomposition for a group of images inside a directory}
\usage{
rgb_decomposition(
  subdirectory,
  extension = "jpg",
  Rdata = TRUE,
  recursive = TRUE
)
}
\arguments{
\item{subdirectory}{main directory to search for images}

\item{extension}{images format or extension [default: jpg]}

\item{Rdata}{boolean flag on whether to store the layers as Rdata or CSV}

\item{recursive}{boolean flag on whether to recursively search for images}
}
\description{
Generate an RGB decomposition for a group of images inside a directory
}
\examples{
\dontrun{
test_data <- data.frame(name = c("R", "G", "B"), values = c(2, 2, 2))
RGB <- c("red", "green", "blue")
png("test_plot.png")
barplot(height = test_data$values, names = test_data$name, col = RGB)
dev.off()
rgb_decomposition(".", "png", recursive = FALSE)
}
}
