% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipa.R
\name{rm_background}
\alias{rm_background}
\title{Remove image background}
\usage{
rm_background(image_path, bkg_thr = 0.4, plot = FALSE, ...)
}
\arguments{
\item{image_path}{filename w/o the full path}

\item{bkg_thr}{background threshold, any pixel below this value will be set
to zero (black) and one (white) in the alpha layer, to create a
transparency effect.}

\item{plot}{boolean flag on whether or not to generate a histogram of the
pixel values. This can be used to determine an optimal \code{bkg_thr}}

\item{...}{extra parameters for \code{hist}}
}
\description{
Remove image background
}
\examples{
\dontrun{
    test_data <- data.frame(name = c("R", "G"), values = c(2, 2))
    RG <- c("red", "green")
    png("test_plot.png")
    par(bg = 'black')
    barplot(height = test_data$values, names = test_data$name, col = RG)
    dev.off()
    rm_background("test_plot.png", 0.1)
    rm_background("test_plot.png", 0.1, TRUE)
    rm_background("test_plot.png", 0.1, TRUE, breaks = 10)
}
}
