% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleIPA.R
\name{simpleIPA}
\alias{simpleIPA}
\title{Single function for the whole IPA analysis}
\usage{
simpleIPA(dataset, adducts.matrix, DB, ionisation, ppm, ppm.thr, RTwin,
  isotopes, iso = TRUE, add = TRUE, bio = TRUE, ints = TRUE,
  rt = TRUE, iso.threshold = 1, corr.matrix = NULL,
  corr.thr = 0.75, relation.id = NULL, pk = NULL, delta.bio = 1,
  delta.add = 1, delta.iso = 1, allsampcomp = NULL,
  ratio.toll = 0.8, allsamp = FALSE, no.its = 1100, burn = 100,
  RT.win.post = 3, IDs = NULL, unknown.ppm = NA, RT.pen = 0.5,
  pr.lim = 1e-05, fully.connected = FALSE, ratios = TRUE,
  connection.type = "reactions", connections = NULL, v = T,
  IT = 500)
}
\arguments{
\item{dataset}{A matrix containing the measured data, organized in 3 colums: mz, RT and Int}

\item{adducts.matrix}{A matrix contains information about all possible adducts.
Columns are: KEGG.id (e.g. 'C00002'), adduct (e.g. 'M+H'), RT (previously known
retention time for the compound, NA if unknown), formula (e.g. C10H17N5O13P3),
theoretical mz (e.g. 508.003570124), charge (e.g. -1) mono ('mono'
if monoisotopical form, 'iso' otherwise)}

\item{DB}{The database used in a dataframe format. It has to be organized in the same fashion of the one included with the package}

\item{ionisation}{A character indicating the ionisation mode of the experiment, either 'positive' or 'negative'}

\item{ppm}{A number indicating the instrument accuracy to be considered}

\item{ppm.thr}{A numerical value indicating the maximum accuracy value to be considered}

\item{RTwin}{A numerical value indicating the maximum difference allowed
between measured RT of different isotopes}

\item{isotopes}{A matrix containing infomation about isotopes}

\item{iso}{A logical parameter. If true isotope connections are used}

\item{add}{A logical parameter. If true adducts connections are used}

\item{bio}{A logical parameter. If true biochemical connections are used}

\item{ints}{A logical parameter. If true intensities are used}

\item{rt}{A logical parameter. If true RTs are used for filtering connections}

\item{iso.threshold}{A numerical value indicating the probability below which
isotope peaks can be omitted}

\item{corr.matrix}{A matrix containing the correlation values between peaks}

\item{corr.thr}{A numerical value expressing the treshold used to consider the correlations significant}

\item{relation.id}{A vector containg class labels of the previously grouped peaks}

\item{pk}{A vector of length nrow(adducts.matrix). Values between 0 and 1 expressing initial confidence of the presence of each formula}

\item{delta.bio}{A number expressing the confidence on the information encoded in Bio
(smaller the value higher the confidence)}

\item{delta.add}{A number expressing the confidence on the information encoded in Add
(smaller the value higher the confidence)}

\item{delta.iso}{A number expressing the confidence on the information encoded in Iso
(smaller the value higher the confidence)}

\item{allsampcomp}{A matrix containing all the assignments for each iteration of the Gibbs sampler in a previus run}

\item{ratio.toll}{A numerical value expressing the minimum accepeted ratio between thereoretical
and observed intensity ratios between isotopes (default 0.8)}

\item{allsamp}{A logical value indicating if the output should contain the allsampcomp object}

\item{no.its}{A numerical value indicating the number of iterations the to be performed by the Gibbs sampler (default 1100)}

\item{burn}{A numerical value indicating the number of initial iterations to be ignored when computing the posterior
probabilities (default 100)}

\item{RT.win.post}{A numerical value expressing the maximum Retention time difference allowed between adducts/isotopes}

\item{IDs}{name of each feature in the initial dataset}

\item{unknown.ppm}{The ppm number to be assigned to unknown (default NA - No unknown is considered)}

\item{RT.pen}{A numerical value indicating the multiplicative factor used when the measured retention time is outside of the range reported in the database}

\item{pr.lim}{A number inidicating the lowest probability value allowed (default 1e-05)}

\item{fully.connected}{A logical value if TRUE all adducts will be considered connected with each other, if FALSE (default) all
the adducts are only connected to the main adduct}

\item{ratios}{A logical value indicating if the matrix should contain the expected intensity ratios (default TRUE)}

\item{connection.type}{if equal to "reactions" (default) considers the reactions reported in the database. If it is equal to "formulas", the function consider the list of possible reactions reported in connections}

\item{connections}{if connection.type=="fomulas", this vector contains a list of formulas that are likely to be added or substracted by an enzymatic reaction. If it is equal to NULL the default list of formulas is used}

\item{v}{A logical value indicating if the progress will be shown (default TRUE)}

\item{IT}{A number inticating after how many iteration an update should be shown (default 120)}
}
\value{
A list containing the Hits, Prior, Post and parsed results
}
\description{
Wrapper summmarizing the whole IPA analysis in one single function. USE IT CAREFULLY!
}
\seealso{
find.Hits compute.Priors IPAposteriors
}
\author{
Francesco Del Carratore \email{francescodc87@gmail.com}
}
