% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{soft_fill}
\alias{soft_fill}
\title{soft imputation.}
\usage{
soft_fill(data, fits)
}
\arguments{
\item{data}{object to be imputed}

\item{fits}{an object returned from \code{\link[=soft_fit]{soft_fit()}}}
}
\value{
a \link[tibble:tibble-package]{tibble} comprising the
multiply imputed data, which are identified by the \code{.impute}
column
}
\description{
an adaptation of the soft impute algorithm for
multiply imputed data.
}
\note{
see \link[softImpute:softImpute]{softimpute} for a more
descriptive summary of the \code{softImpute} algorithm.
}
\references{
Rahul Mazumder, Trevor Hastie and Rob Tibshirani (2010)
Spectral Regularization Algorithms for Learning Large Incomplete
Matrices, http://www.stanford.edu/~hastie/Papers/mazumder10a.pdf
\emph{Journal of Machine Learning Research} 11 (2010) 2287-2322
}
\author{
Trevor Hastie, Rahul Mazumder
}
