% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{soft_mash}
\alias{soft_mash}
\title{Mash a soft mash}
\usage{
soft_mash(malt, n_impute = 10, step_size = 1, scale_data = TRUE,
  scale_iter = 20, scale_lambda = 0.95, lambda_sequence = NULL)
}
\arguments{
\item{malt}{a \code{soft_brew} object that has been malted
(see \link{soft_malt}).}

\item{n_impute}{An integer indicating the number of imputed
datasets to create.}

\item{step_size}{An integer indicating the number by which to increase
the maximum rank of the softImpute solution after each iteration.}

\item{scale_data}{\code{TRUE} or \code{FALSE}. If \code{TRUE}, the
\code{biScale} algorithm will be applied to the training
data prior to applying the \code{softImpute} algorithm. It
is recommended that this argument be set to \code{TRUE}}

\item{scale_iter}{An integer indicating the maximum number
of iterations that will be applied in the \code{biScale}
algorithm. It is recommended that this argument be managed
to ensure convergence.}

\item{scale_lambda}{A positive double less than 1. This constant
scales the maximum value of \code{lambda} used to fit the
\code{softImpute} function. Larger values of \code{lambda}
correspond to lower maximum rank solutions.}

\item{lambda_sequence}{A sequence of \code{lambda} values. If
specified, these \code{lambda} values will be used instead of
the \code{lambda} values that \code{soft_mash()} automatically identifies.}
}
\description{
The \code{softImpute} algorithm has a number of
inputs that affect the imputation procedure. Parameters
set in this step of the \code{brew} process include the
progression of the maximum rank of \code{softImpute} solutions,
whether or not to apply the \code{biScale} function (see notes),
and additional parameters that give the user some control
over \code{lambda}. See notes for links to the original
\code{softImpute} functions.
}
\note{
see \link[softImpute:softImpute]{softimpute} for a more
descriptive summary of the \code{softImpute} algorithm. See
\link[softImpute:biScale]{biScale} for a more descriptive
summary of the \code{biScale} algorithm.
}
\section{Overview}{

After malting a \code{brew}, the next step is to pair the
malted wheat with hot water and mash it. In the \code{ipa} process,
mashing is where an imputation method is paired with relevant
hyper-parameters.
}

\references{
Rahul Mazumder, Trevor Hastie and Rob Tibshirani (2010)
Spectral Regularization Algorithms for Learning Large Incomplete
Matrices, http://www.stanford.edu/~hastie/Papers/mazumder10a.pdf
\emph{Journal of Machine Learning Research} 11 (2010) 2287-2322
}
\author{
Trevor Hastie, Rahul Mazumder
}
