% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spice.R
\name{spice}
\alias{spice}
\title{Spice a brew}
\usage{
spice(brew, with = NULL, ...)
}
\arguments{
\item{brew}{an \code{ipa_brew} object.}

\item{with}{a helper function for mashing brews.
See \link{spicer_nbrs} and \link{spicer_soft})}

\item{...}{additional arguments for specific brew flavors.}
}
\value{
an \code{ipa_brew} object with values added to \code{pars}.
}
\description{
When an \code{ipa_brew} is spiced, tuning parameters
are designated and saved in the \code{pars} component of the \code{brew}.
}
\examples{

data <- data.frame(
  x1 = 1:10,
  x2 = 10:1,
  x3 = rnorm(10),
  outcome = 11 + runif(10)
)

data[1:2, 1:2] = NA

knn_brew <- brew(data, outcome = outcome, flavor = 'kneighbors')
sft_brew <- brew(data, outcome = outcome, flavor = 'softImpute')

# these two calls are equivalent
spicy_knn <- spice(knn_brew, with = spicer_nbrs(k_neighbors = c(3, 5)))
spicy_knn <- spice(knn_brew, k_neighbors = c(3, 5))

# these two calls are equivalent
spicy_sft <- spice(sft_brew, with = spicer_soft(grid = TRUE))
spicy_sft <- spice(sft_brew, grid = TRUE)


}
