% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spice.R
\name{spicer_nbrs}
\alias{spicer_nbrs}
\title{Neighbor's spices}
\usage{
spicer_nbrs(neighbors = NULL, aggr_neighbors = TRUE)
}
\arguments{
\item{neighbors}{the number of neighbors used for imputation.
One imputed dataset is created for each value given. For
example, to make two imputed datasets using 1 and 2
nearest neighbors, one would write \code{neighbors = c(1,2)}}

\item{aggr_neighbors}{(\code{TRUE} / \code{FALSE}) whether neighbor's
values should be aggregated or sampled. If \code{TRUE}, then
the mean/mode of neighboring values will be used. Otherwise,
one neighbor's value will be sampled at random.}
}
\value{
a list with parameter values that can be passed
directly into \code{kneighbors_brew} objects via \link{spice}.
}
\description{
It can be a little overwhelming to remember which sets of
parameters go with each \code{ipa_brew} flavor, so just add a dash
from a handy \code{spicer} function and get on with your \code{brew}.
}
\concept{spices}
