% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spice.R
\name{spicer_rngr}
\alias{spicer_rngr}
\title{Ranger's spices}
\usage{
spicer_rngr(min_node_sizes = NULL, pmm_donor_sizes = 0L)
}
\arguments{
\item{min_node_sizes}{Minimal node size of a leaf in a decision tree.
Increasing the minimal node size reduces the chance of overfitting.}

\item{pmm_donor_sizes}{Number of candidate non-missing values to
sample from in the predictive mean matching step.
Set \code{pmm_donor_sizes = 0} to skip this step.}
}
\value{
a list with parameter values that can be passed
directly into \code{missRanger_brew} objects via \link{spice}.
}
\description{
It can be a little overwhelming to remember which sets of
parameters go with each \code{ipa_brew} flavor, so just add a dash
from a handy \code{spicer} function and get on with your \code{brew}.
}
\concept{spices}
