% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_hot.R
\name{spread_cats}
\alias{spread_cats}
\alias{gather_cats}
\title{one-hot coding}
\usage{
spread_cats(data, factor_variables = NULL, ...)

gather_cats(data, factor_levels)
}
\arguments{
\item{data}{data with categorical variables (i.e., factors)
that need to be spread or gathered.}

\item{...}{Arguments passed on to \code{mltools::one_hot}
\describe{
  \item{sparsifyNAs}{Should NAs be converted to 0s?}
  \item{naCols}{Should columns be generated to indicate the present of NAs? Will only apply to factor columns with at least one NA}
  \item{dropCols}{Should the resulting data.table exclude the original columns which are one-hot-encoded?}
}}

\item{factor_levels}{This parameter is only relevant for
\code{gather_cats}. A named list of factor levels, with each
name corresponding to the column in the data that the
factor levels describe.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Many modeling functions expect matrix input
with factor levels one-hot encoded. \code{spread_cats}
will one-hot encode any factor or character variable
in \code{data} and return a one-hot encoded \code{tibble}.
Alternatively, \code{gather_cats} will apply the inverse
operation and convert one-hot encoded columns back
into factors.
}
\note{
factor levels are stored as an attribute
in the output of \code{spread_cats}. If this attribute
is intact when \code{gather_cats} is applied to the
one-hot encoded data, it will not be necessary
to specify \code{factor_levels}
}
