% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_data.R
\name{stack_data}
\alias{stack_data}
\alias{list_data}
\title{Data Stacks & Lists}
\usage{
stack_data(data, keep_ID = FALSE)

list_data(data, keep_ID = FALSE)
}
\arguments{
\item{data}{(\code{list}) This argument should be a list of
imputed dataframes, each having the same number of rows and columns.}

\item{keep_ID}{(\code{TRUE / FALSE}) Should the returned object include
a column named \code{._ID_.} that indicates which observations in the
stacked/listed imputed data belong to the same observation in the
original unimputed data?}
}
\description{
stacked dataframes comprise multiply imputed
dataframes that are stacked on top of one another.
A single models can be fit to stacked data, and the
additional diversity in the data can sometimes reduce
the model's prediction error. Data lists are similar,
to data stacks, but data frames in a list are kept
separate and one model is fit for each dataframe.

These functions create ID variables and attach them
to the output data.
}
\examples{
data_list <- list(
  df1 = data.frame(x = 1:3, y = letters[1:3], stringsAsFactors=FALSE),
  df2 = data.frame(x = 2:4, y = letters[1:3], stringsAsFactors=FALSE)
)

stack_data(data_list, keep_ID = TRUE)

list_data(data_list, keep_ID = TRUE)
}
