% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stir.R
\name{stir}
\alias{stir}
\title{Stir your brew}
\usage{
stir(brew, timer = FALSE)
}
\arguments{
\item{brew}{an \code{ipa_brew} object.}

\item{timer}{a logical value. If \code{TRUE}, then the amount of time it takes
to fit the imputation models will be tracked and saved as an attribute
of the resulting \code{ipa_brew} object.}
}
\value{
an \code{ipa_brew} object with imputed values added to the \code{wort}
}
\description{
After setting your primary and secondary imputation parameters
(i.e., running \link{spice} and \link{mash}), you're ready to fit the
imputation model(s) that will impute missing values in your
training data. This is what the \code{stir} function does.
}
\examples{

x1 = rnorm(100)
x2 = rnorm(100) + x1
x3 = rnorm(100) + x1 + x2

outcome = 0.5 * (x1 - x2 + x3)

n_miss = 10
x1[1:n_miss] <- NA

data <- data.frame(x1=x1, x2=x2, x3=x3, outcome=outcome)

sft_brew <- brew_soft(data, outcome=outcome, bind_miss = FALSE)
sft_brew <- mash(sft_brew, with = masher_soft(bs = TRUE))
sft_brew <- stir(sft_brew, timer = TRUE)
sft_brew

}
