% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_impute_handlers.R
\name{test_stkr}
\alias{test_stkr}
\title{Impute test data}
\usage{
test_stkr(data, dbl_impute = TRUE)
}
\arguments{
\item{data}{data to be imputed.}

\item{dbl_impute}{(\code{TRUE} / \code{FALSE}) if \code{TRUE}, then each imputed
training set is stacked on top of the testing data, separately, prior
to fitting each imputation model for the testing data. If \code{FALSE},
then the original data is stacked on top of the testing data, and
each imputation model is fitted to the same stacked dataset.}
}
\value{
An object of class 'dbl_imputer'
}
\description{
Some missing data strategies (e.g., soft imputation) are not
designed to impute missing values in testing data. However, the
\code{softImpute} strategy may still be applied to the testing data
using the same fit parameters as in the imputation model for the
training data. In addition, to enhance the accuracy of test imputation
models, the training data are stacked on top of the testing data
prior to fitting the models.
}
\note{
This function should only be applied within the \code{ferment}
function. Outside of this context, \code{test_stkr()} has no purpose.
see ?ferment and vignettes for more detailed examples
}
