% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updating_DBs.R
\name{update_database}
\alias{update_database}
\title{Function used to update your database}
\usage{
update_database(DB, all.adducts.matrix.POS, all.adducts.matrix.NEG, update,
  adducts, isotopes)
}
\arguments{
\item{DB}{The database used in a dataframe format. It has to be organized in the same fashion of the one included with the package}

\item{all.adducts.matrix.POS}{A matrix contains information about all possible adducts.
Columns are: KEGG.id (e.g. 'C00002'), adduct (e.g. 'M+H'), RT (previously known
retention time for the compound, NA if unknown), formula (e.g. C10H17N5O13P3),
theoretical mz (e.g. 508.003570124), charge (e.g. 1) mono ('mono'
if monoisotopical form, 'iso' otherwise)}

\item{all.adducts.matrix.NEG}{A matrix contains information about all possible adducts.
Columns are: KEGG.id (e.g. 'C00002'), adduct (e.g. 'M-H'), RT (previously known
retention time for the compound, NA if unknown), formula (e.g. C10H17N5O13P3),
theoretical mz (e.g. 508.003570124), charge (e.g. -1) mono ('mono'
if monoisotopical form, 'iso' otherwise)}

\item{update}{A dataframe containing the new information}

\item{adducts}{A dataframe containing adducts information as the one included in the enviPat package}

\item{isotopes}{A dataframe containing isotopes information as the one included in the enviPat package}
}
\value{
Updated database
}
\description{
This function updates the database according to new information
}
\author{
Francesco Del Carratore \email{francescodc87@gmail.com}
}
