% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brew.R
\name{verbose_on}
\alias{verbose_on}
\alias{verbose_off}
\alias{get_verbosity}
\title{Noisier brewing}
\usage{
verbose_on(brew, level)

verbose_off(brew)

get_verbosity(brew)
}
\arguments{
\item{brew}{an \code{ipa_brew} object (see \link{brew}).}

\item{level}{The level of verbosity.}
}
\value{
a noisier \code{ipa_brew} object with
an adjusted \code{verbose} attribute value.
}
\description{
Sometimes you want to see text output to make sure
your brew is brewing the way you want it to. Use this function
to make your brew tell you what's going on at all subsequent
stages.
}
