% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{bottle}
\alias{bottle}
\title{Bottle a brew}
\usage{
bottle(brew, flavor = c("tibble", "matrix"), drop_fit = TRUE)
}
\arguments{
\item{brew}{a \code{brew} object that has been passed through all
the brewing steps: \code{malt} (i.e., \link{malt_nbrs}, \link{malt_soft}),
\code{mash} (i.e., \link{mash_nbrs}, \link{mash_soft}),
\link{boil}, and \link{ferment}.}

\item{flavor}{a character value indicating what composition
the training and testing data will be returned as.}
}
\description{
Bottle a brew
}
