% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{ferment}
\alias{ferment}
\alias{ferment.soft_brew}
\alias{ferment.neighbors_brew}
\title{Ferment a brew}
\usage{
ferment(brew, new_data = NULL, col_name = "testing",
  control = list())

\method{ferment}{soft_brew}(brew, new_data = NULL,
  col_name = "testing", control = list())

\method{ferment}{neighbors_brew}(brew, new_data = NULL,
  col_name = "testing", control = list())
}
\arguments{
\item{brew}{a \code{brew} object that has been boiled (see \link{boil}).}

\item{new_data}{testing data (i.e., data that will be used to
test a predictive model) with missing values.}

\item{col_name}{a character value indicating what the name of
the imputed datasets will be in the \code{wort} (see notes).}

\item{neighbors}{an integer value specifying the number of
nearest neighbors to identify when imputing testing data.
(see notes.)}

\item{impute_index}{the index of the imputed training
dataset that will be used to identify nearest neighbors
in testing data. (see notes)}
}
\description{
Ferment a brew
}
\section{Methods (by class)}{
\itemize{
\item \code{soft_brew}: impute missing values for \code{soft_brew} objects.

\item \code{neighbors_brew}: impute missing values for \code{neighbors_brew} objects.
}}

\note{
That the \code{wort} is a component of \code{brew} objects that
contains imputed datasets, models used to impute those datasets,
and the corresponding hyper-parameters of those models.

The \code{neighbors} and \code{impute_index} arguments are
necessary for imputation methods that are not compatible
with imputation of testing data.
\itemize{
\item \link[softImpute:softImpute]{softimpute} imputation method
can only be applied too impute training data. So, when
testing data need to be imputed, \code{ipa} will look for
nearest neighbors of testing observations in the \emph{imputed}
training data. And because there may be more than one
imputed training data set, \code{impute_index} allows the user
to specify which one is used to find nearest neighbors
for testing observations.
}

The \code{neighbors} and \code{impute_index} arguments are
applicable (though not necessary) for k-nearest-neighbor
imputation objects (i.e., \code{neighbors_brew} objects).
}
