% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipa.R
\name{find_area}
\alias{find_area}
\title{Find continuos group of pixels with a tolerance of \code{px_tol} pixels}
\usage{
find_area(img, start = c(1, 1), px_tol = 20)
}
\arguments{
\item{img}{image as cimg file}

\item{start}{starting point: \code{c(x0, y0)}}

\item{px_tol}{number of non-continuos pixels to accept}
}
\value{
blobs containg adjacent groups of non-zero pixels
}
\description{
Find continuos group of pixels with a tolerance of \code{px_tol} pixels
}
