% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{gen_simdata}
\alias{gen_simdata}
\title{a simple framework to simulate simple dataframes}
\usage{
gen_simdata(problem_type = c("regression", "classification", "survival"),
  ncov = 3, nint = 2, degree = 3, rho = 1/2, corstr = c("AR1",
  "CS"), nobs = 10000, error_sd = 1/2, prevalence = NULL,
  split_prop = 1/2, miss_pattern = "mar", npatterns = 10,
  trn_miss_prop = 1/2, tst_miss_prop = 0)
}
\arguments{
\item{problem_type}{A character value indicating the problem type to
simulate data for. Valid options are 'regression', 'classification',
or 'survival'.}

\item{ncov}{the number of main effects used to generate an outcome variable}

\item{nint}{the number of interaction effects used to generate an outcome
variable.}

\item{degree}{the degree of each predictor variable's relationship to
the outcome. For example, \code{degree = 2} makes the relationship between
each predictor variable and the outcome quadratic.}

\item{rho}{the correlation constant among predictors in the X matrix.}

\item{corstr}{The correlation structure among predictors in the X matrix.}

\item{nobs}{the total number of observations in the simulated data.}

\item{error_sd}{the standard deviation of error applied when generating
outcome values.}

\item{prevalence}{(this is only relevant for classification problems).
The prevalence of the outcome.}

\item{split_prop}{the proportion of data that will be randomly assigned
to the training dataset.}

\item{miss_pattern}{A character value indicating what type of missingness
pattern to apply. Valid options are 'mcar', 'mar', and 'mnar'.
mcar = missing completely at random, mar = missing at random, and
mnar = missing not at random.}

\item{npatterns}{The number of missing patterns used to ampute data.}

\item{trn_miss_prop}{The proportion of data in the training set that will
be set to missing.}

\item{tst_miss_prop}{The proportion of data in the testing set that will
be set to missing.}
}
\description{
Simulated data allow analysts to conduct controlled experiments
using specific parameters to generate data. This function provides
a general framework to simulate missing data under different patterns:
}
\details{
\enumerate{
\item missing completely at random (MCAR): missingness occurs at random,
independent of all variables
\item missing at random (MAR): missingness occurs at random, conditional
on measured variables in the observed data.
\item missing not at random (MNAR) missingess occurs non-randomly and is
dependent on unmeasured variables.
}

All simulated predictor variables are numeric. Regression coefficients
are generated randomly (all values are between -1 and 1).
}
\examples{

regr = gen_simdata(problem_type = 'regression',
  ncov = 3, nint = 2, degree = 3, nobs = 2000,
  tst_miss_prop = 0)

clsf = gen_simdata(problem_type = 'classification',
  ncov = 3, nint = 2, degree = 3, nobs = 2000)


surv = gen_simdata(problem_type = 'survival',
  ncov = 3, nint = 2, degree = 3, nobs = 2000)
}
