% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipa.R
\docType{package}
\name{ipa-package}
\alias{ipa}
\alias{ipa-package}
\title{imputation for predictive analytics - ipa.}
\description{
ipa provides functions to help you design and use effective
strategies to handle missing data.
}
\details{
To learn more about ipa, start with the vignettes:
\code{browseVignettes(package = "ipa")}
}
\author{
\strong{Maintainer}: Byron Jaeger \email{bcjaeger@uab.edu} (\href{https://orcid.org/0000-0001-7399-2299}{ORCID})

}
