% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipa.R
\encoding{UTF-8}
\name{ipa}
\alias{ipa}
\alias{sampa}
\alias{xsampa}
\alias{arpa}
\alias{arpabet}
\alias{cmu}
\title{Convert between IPA, X-SAMPA and ARPABET (CMU)}
\usage{
ipa(x, to = c("xsampa", "arpabet"))

sampa(x, to = c("ipa", "arpabet"))

xsampa(x, to = c("ipa", "arpabet"))

arpa(x, to = c("ipa", "xsampa"))

arpabet(x, to = c("ipa", "xsampa"))

cmu(x, to = c("ipa", "xsampa"))
}
\arguments{
\item{x}{A phonetic character vector in IPA (for \code{ipa()}),
X-SAMPA (for \code{sampa()}) or ARPABET (for \code{arpa()}).}

\item{to}{The phonetic alphabet to convert to.
Defaults to \code{"ipa"} for \code{sampa()} and \code{arpa()} and \code{"xsampa"} for \code{ipa()}.
\code{"sampa"} is an accepted alias for \code{"xsampa"}.
\code{"arpa"} and \code{"cmu"} are accepted aliases for \code{"arpabet"}.}
}
\value{
A character vector the same length as \code{x}
}
\description{
\itemize{
\item \code{ipa()} converts from IPA.
\item \code{sampa()} and \code{xsampa()} convert from X-SAMPA.
\item \code{arpa()}, \code{arpabet()} and \code{cmu()} convert from ARPABET (used by the CMU
Pronouncing Dictionary).
}
}
\examples{
sampa("aI pi: \"eI")
sampa(c("\"nom.b4e", "nO~bR"))
sampa("Ar\\\\pV", "arpa")

arpa("AA R P AH")
arpa("AA R P AH", "xsampa")

# `ipa()` supports raw Unicode input or escaped Unicode sequences (\uxxxx),
# but raw Unicode is not supported in R package documentation,
# hence the ugly examples

ipa("a\u026a pi\u02d0 \'e\u026a")
ipa(c("\'nom.b\u027ee", "n\u0254\u0303b\u0281"))
ipa("\u0251\u0279p\u028c", "arpa")
}
