% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipa.R
\name{ipa}
\alias{ipa}
\alias{sampa}
\alias{xsampa}
\title{Convert between IPA and X-SAMPA}
\usage{
ipa(x, to = c("xsampa", "ipa"))

sampa(x, to = c("ipa", "xsampa"))

xsampa(x, to = c("ipa", "xsampa"))
}
\arguments{
\item{x}{A phonetic character vector in (X-)SAMPA (for \code{sampa()}) or IPA (for
\code{ipa()})}

\item{to}{The phonetic alphabet to convert to.
Defaults to IPA for \code{sampa()} and X-SAMPA for \code{ipa()}.}
}
\value{
A character vector the same langth as \code{x}
}
\description{
\code{sampa()} converts (X-)SAMPA to IPA. \code{ipa()} converts IPA to X-SAMPA.
}
\examples{
sampa("aI pi: \"eI")
sampa(c("nomb4e", "nO~bR"))

ipa("aÉª piË ËˆeÉª")
ipa(c("nombÉ¾e", "nÉ”ÌƒbÊ"))
}
