% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipa.R
\name{ipa}
\alias{ipa}
\alias{sampa}
\alias{arpa}
\title{Convert between IPA and X-SAMPA}
\usage{
ipa(x, to = c("xsampa", "arpabet"))

sampa(x, to = c("ipa", "arpabet"))

arpa(x, to = c("ipa", "xsampa"))
}
\arguments{
\item{x}{A phonetic character vector in IPA (for \code{ipa()}),
X-SAMPA (for \code{sampa()}) or ARPABET (for \code{arpa()}).}

\item{to}{The phonetic alphabet to convert to.
Defaults to \code{"ipa"} for \code{sampa()} and \code{arpa()} and \code{"xsampa"} for \code{ipa()}.}
}
\value{
A character vector the same length as \code{x}
}
\description{
\itemize{
\item \code{ipa()} converts from IPA.
\item \code{sampa()} converts from X-SAMPA.
\item \code{arpa()} converts from ARPABET (used by the CMU Pronouncing Dictionary).
}
}
\examples{
sampa("aI pi: \"eI")
sampa(c("\"nom.b4e", "nO~bR"))

ipa("aÉª piË ËˆeÉª")
ipa(c("Ëˆnom.bÉ¾e", "nÉ”ÌƒbÊ"))

arpa("AA R P AH")
arpa("AA R P AH", "xsampa")
sampa("Ar\\\\pV", "arpa")
ipa("É‘É¹pÊŒ", "arpa")
}
