% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mash.R
\name{masher_nbrs}
\alias{masher_nbrs}
\title{Neighbor's masher}
\usage{
masher_nbrs(
  eps = 1e-08,
  nthread = getOption("gd_num_thread"),
  fun_aggr_ctns = mean,
  fun_aggr_intg = medn_est,
  fun_aggr_catg = mode_est
)
}
\arguments{
\item{eps}{A numeric value > 0. Computed numbers (variable ranges)
smaller than \code{eps} are treated as zero}

\item{nthread}{Number of threads to use for parallelization.
By default, for a dual-core machine, 2 threads are used.
For any other machine n-1 cores are used so your machine doesn't
freeze during a big computation. The maximum number of threads
are determined using \code{omp_get_max_threads} at C level.}
}
\value{
a list with \link[gower:gower_topn]{gower_topn} input values
that can be passed directly into \code{kneighbors_brew} objects via \link{mash}.
}
\description{
It can be a little overwhelming to remember which sets of
parameters go with each \code{ipa_brew} flavor, so just pair your
flavor with its \code{masher} function and get on with your \code{brew}.
}
