% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mash.R
\name{masher_soft}
\alias{masher_soft}
\title{Soft masher}
\usage{
masher_soft(
  bs = TRUE,
  bs_maxit = 20,
  bs_thresh = 1e-09,
  bs_row.center = FALSE,
  bs_col.center = TRUE,
  bs_row.scale = FALSE,
  bs_col.scale = TRUE,
  si_type = "als",
  si_thresh = 1e-05,
  si_maxit = 100,
  si_final.svd = TRUE
)
}
\arguments{
\item{bs}{a logical value. If \code{TRUE}, \code{\link[softImpute:biScale]{softImpute::biScale()}} is applied
to \code{data_ref} or \code{rbind(data_ref, data_new)} prior to fitting \code{softImpute}
models.}

\item{bs_maxit}{an integer indicating the maximum number of iterations
for the \code{biScale} algorithm.}

\item{bs_thresh}{convergence threshold for the \code{biScale} algorithm.}

\item{bs_row.center}{a logical value. If \code{TRUE}, row centering
will be performed. If \code{FALSE} (default), then nothing is done.}

\item{bs_col.center}{a logical value. If \code{TRUE} (default), column centering
will be performed. If \code{FALSE}, then nothing is done.}

\item{bs_row.scale}{a logical value. If \code{TRUE}, row scaling
will be performed. If \code{FALSE} (default), then nothing is done.}

\item{bs_col.scale}{a logical value. If \code{TRUE} (default), column scaling
will be performed. If \code{FALSE}, then nothing is done.}

\item{si_type}{two algorithms are implemented, type="svd" or the default
type="als". The "svd" algorithm repeatedly computes the svd of the
completed matrix, and soft thresholds its singular values. Each new
soft-thresholded svd is used to re-impute the missing entries. For
large matrices of class "Incomplete", the svd is achieved by an
efficient form of alternating orthogonal ridge regression. The
"als" algorithm uses this same alternating ridge regression, but
updates the imputation at each step, leading to quite substantial
speedups in some cases. The "als" approach does not currently
have the same theoretical convergence guarantees as the
"svd" approach.}

\item{si_thresh}{convergence threshold for the \code{softImpute} algorithm,
measured as the relative change in the Frobenius norm between two
successive estimates.}

\item{si_maxit}{maximum number of iterations for the \code{softImpute}
algorithm.}

\item{si_final.svd}{only applicable to \code{si_type = "als"}. The alternating
ridge-regressions do not lead to exact zeros. With the default
\code{final.svd = TRUE}, at the final iteration, a one step unregularized
iteration is performed, followed by soft-thresholding of the
singular values, leading to hard zeros.}
}
\value{
a list with input values that can be passed directly into
\link{mash}, e.g \code{mash(brew, with = masher_nbrs())} for a neighbors brew or
\code{mash(brew, with = masher_soft())} for a soft brew.
}
\description{
If you use Rstudio, the \code{masher} and \code{spicer} functions can help
remind you which parameters go along with which \code{ipa_brew} flavor.
The basic idea is to write \verb{spice(brew, with = spicer_<flavor>())}
and \verb{mash(brew, with = masher_<flavor>())}. Hitting the tab key with
your curser inside the parentheses of \code{masher_flavor()}will create a
drop-down menu that shows a list of the arguments that go along with
your brew's flavor.

If you have no trouble remembering the parameters that go along
with your brew's flavor, or if you just want your code to be more concise,
you don't have to use the \code{with} argument. Instead, you can just
specify parameter values directly using the \code{...} argument in the \code{mash}
and \code{spice} functions. In the examples below, both approaches are shown.
}
\examples{

x1 = rnorm(100)
x2 = rnorm(100) + x1
x3 = rnorm(100) + x1 + x2

outcome = 0.5 * (x1 - x2 + x3)

data <- data.frame(x1=x1, x2=x2, x3=x3, outcome=outcome)

n_miss = 10

data[1:n_miss,'x1'] = NA
sft_brew <- brew_soft(data, outcome=outcome, bind_miss = FALSE)

# these two calls are equivalent
mash(sft_brew, with = masher_soft(bs = FALSE))
mash(sft_brew, bs = FALSE)

knn_brew <- brew_nbrs(data, outcome=outcome, bind_miss = TRUE) \%>\%

# these two calls are equivalent
mash(knn_brew, with = masher_nbrs(fun_aggr_ctns = median))
mash(knn_brew, fun_aggr_ctns = median)

}
