% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgboost.R
\name{mgb_surv_prob}
\alias{mgb_surv_prob}
\title{Predicted values from \code{midy} boosters.}
\usage{
mgb_surv_prob(object, new_data, eval_times, new_n_impute, new_miss_strat)
}
\arguments{
\item{object}{an object of class \code{xgb.Booster}.}

\item{eval_times}{(numeric) a vector of times for which model
predictions are computed.}
}
\description{
Computing survival probabilities with gradient
boosted decision tree ensembles.
}
