% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spice.R
\name{spice}
\alias{spice}
\title{Spice a brew}
\usage{
spice(brew, with = NULL, ...)
}
\arguments{
\item{brew}{an \code{ipa_brew} object.}

\item{with}{the output of a helper function for spicing brews.
The helper functions are \link{spicer_nbrs} and \link{spicer_soft}}

\item{...}{additional arguments for specific brew flavors.}
}
\value{
a spiced \code{ipa_brew}, with imputation
parameters just how you like them.
}
\description{
When an \code{ipa_brew} is spiced, tuning parameters
are designated and saved in the \code{pars} component of the \code{brew}.
}
\examples{

data <- data.frame(
  x1 = 1:10,
  x2 = 10:1,
  x3 = rnorm(10),
  outcome = 11 + runif(10)
)

data[1:2, 1:2] = NA

knn_brew <- brew(data, outcome = outcome, flavor = 'kneighbors')
spicy_knn <- spice(knn_brew, k_neighbors = c(3, 5), aggregate = TRUE)

sft_brew <- brew(data, outcome = outcome, flavor = 'softImpute')
spicy_sft <- spice(sft_brew, with = spicer_soft(grid = TRUE))

}
