% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spice.R
\name{spicer_nbrs}
\alias{spicer_nbrs}
\title{Neighbor's spices}
\usage{
spicer_nbrs(k_neighbors = seq(10), aggregate = TRUE)
}
\arguments{
\item{k_neighbors}{a numeric vector indicating how many neighbors should
be used to impute missing values.}

\item{aggregate}{a logical value. If \code{TRUE}, then neighbors
will be aggregated to generate imputations. If \code{FALSE}, then one
neighbor will be sampled at random to generate a missing value. Using
\code{aggregate = FALSE} can be helpful if you are conducting
multiple imputation.}
}
\value{
a list with parameter values that can be passed
directly into \code{kneighbors_brew} objects via \link{spice}.
}
\description{
It can be a little overwhelming to remember which sets of
parameters go with each \code{ipa_brew} flavor, so just add a dash
from a handy \code{spicer} function and get on with your \code{brew}.
}
\concept{spices}
