% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spice.R
\name{spicer_soft}
\alias{spicer_soft}
\title{Soft spices}
\usage{
spicer_soft(n_impute = NULL, step_size = 1L)
}
\arguments{
\item{n_impute}{An integer indicating the number of imputed
datasets to create.}

\item{step_size}{An integer indicating the number by which to increase
the maximum rank of the softImpute solution after each iteration.}
}
\value{
a list with parameter values that can be passed
directly into \code{missRanger_brew} objects via \link{spice}.
}
\description{
It can be a little overwhelming to remember which sets of
parameters go with each \code{ipa_brew} flavor, so just add a dash
from a handy \code{spicer} function and get on with your \code{brew}.
}
\concept{spices}
