% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_data.R
\name{stack_data}
\alias{stack_data}
\alias{list_data}
\title{Prepare data stacks/lists}
\usage{
stack_data(data)

list_data(data)
}
\arguments{
\item{data}{(a list) This argument should be a list of
imputed dataframes, each having the same number of rows and columns.}
}
\description{
stacked dataframes comprise multiply imputed
dataframes that are stacked on top of one another. Imputed values in
these objects are usually a model prediction + random noise.
The additional noise creates diversity among the multiple
dataframes. Data lists are similar, but data frames are kept
separate and one model is fit for each dataframe.

These functions create ID variables and attach classes and attributes
that are used to direct downstream analyses.
}
\examples{
data_list <- list(
  df1 = data.frame(a = 1:3, b = letters[1:3]),
  df2 = data.frame(a = 2:4, b = letters[2:4])
)

stack_data(data_list)

as_mi(data_list)

as_si(data_list[[1]])
}
