% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SyntheticExperiment.R
\docType{SyntheticExperiment}
\name{SyntheticExperiment}
\alias{SyntheticExperiment}
\title{Database}
\format{A dataset with 83 simulated features}
\usage{
data(SyntheticExperiment)
}
\description{
Dataset with the 83 simulated positive features
}
\keyword{datasets}
